backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets<14
mss>=9.0.1
Pillow>=10.2.0
numpy>=1.26.0
python-multipart>=0.0.9
pytesseract>=0.3.10
google-generativeai>=0.3.0
requests>=2.31.0
sounddevice>=0.4.6
faster-whisper>=1.0.0
httpx>=0.27.0
